LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/2d.conformance.requirements.delete-expected.txt

2d.conformance.requirements.delete
window.CanvasRenderingContext2D is Configurable
Defined in "Web IDL" (draft)

Actual output:

PASS window.CanvasRenderingContext2D is Configurable
